# perf: Optimize subscriptions for scans with a limit