@grgbkr 2021-11-05
@arv this is ready for review now.  Let me know if you can think of any other tests to write.

I also don't know too much about subscriptions/scan and indexes .  Are there any considerations around indexes we are missing?